% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nothing.R
\name{do.nothing}
\alias{do.nothing}
\alias{do.nothing.else}
\title{Do nothing, do nothing but nothing.}
\usage{
do.nothing(...)
}
\arguments{
\item{\dots}{Arguments avail not.}
}
\value{
Practice not-doing, and everything will fall into place.
}
\description{
Nothing is done, nothing is left undone.
}
\details{
If nothing is done, then all will be well.
}
\examples{

do.nothing()
\dontrun{
do.nothing.else()
}

}
\author{
Lao Tsu, Aaron A. King
}
\keyword{programming}
