% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tao-package.R
\docType{package}
\name{tao-package}
\alias{tao-package}
\title{That which cannot be packaged.}
\description{
The package that can be described is not the great Tao package.
}
\details{
Without opening your door, you can open your heart to the world.\cr
Without looking out your window, you can see the essence of the Tao.\cr
The more you know, the less you understand.\cr
The Master arrives without leaving, sees the light without looking, achieves without doing a thing.
}
\references{
Lao Tsu, Tao Te Ching.
}
\seealso{
\code{\link{do.nothing}}, \code{\link{do.nothing.else}}
}
\author{
Lao Tsu, Aaron A. King
}
\keyword{programming}
