\name{do.nothing}
\alias{do.nothing}
\alias{do.nothing.else}
\title{Do nothing, do not do nothing.}
\description{
  Nothing is done, nothing is left undone.
}
\usage{
do.nothing(\dots)
do.nothing.else(\dots)
}
\arguments{
  \item{\dots}{Arguments avail not.}
}
\value{
  The softest thing on the universe overcomes the hardest thing in the universe.
  That without substance can enter where there is no room.
  Hence I know the value of non-action.
  Teaching without words and work without doing are understood by very few.
}
\details{If nothing is done, then all will be well.}
\author{Aaron A. King (kingaa at umich dot edu)}
\references{}
\examples{
do.nothing()
\dontrun{
do.nothing.else()
}
}
\keyword{programming}

