\name{do.nothing}
\alias{do.nothing}
\alias{do.nothing.else}
\title{Do nothing, do not do nothing.}
\description{
  Nothing is done, nothing is left undone.
}
\usage{
do.nothing(\dots)
do.nothing.else(\dots)
}
\arguments{
  \item{\dots}{Arguments avail not.}
}
\value{
 Practice not-doing, and everything will fall into place.
}
\details{If nothing is done, then all will be well.}
\author{Lao Tsu, Aaron A. King}
\examples{
do.nothing()
\dontrun{
do.nothing.else()
}
}
\keyword{programming}

