% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nothing.R
\name{do.nothing}
\alias{do.nothing}
\alias{do.nothing.else}
\title{If nothing is done, then all will be well.}
\usage{
do.nothing(...)

do.nothing.else(...)
}
\arguments{
\item{\dots}{Arguments avail not.}
}
\value{
What difference between yes and no? \cr
What difference between success and failure? \cr
Must you value what others value, avoid what others avoid? \cr
How ridiculous! \cr
}
\description{
Nothing is done, nothing is left undone.
}
\examples{
do.nothing()

\dontrun{
  do.nothing.else()
}
}
\author{
Lao Tsu, Aaron A. King
}
\keyword{programming}
