% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tao-package.R
\docType{package}
\name{tao-package}
\alias{tao}
\alias{tao-package}
\title{That which cannot be packaged.}
\description{
The package that can be described is not the great Tao package.
}
\details{
The package that can be documented is not the eternal package \cr
The procedure that terminates is not the eternal procedure \cr
The undocumented is the origin of feature and bug \cr
The documentation is the mother of limitless confusion \cr
Thus, constantly without desire, one observes the runtime errors \cr
Constantly with desire, one observes the progress bar \cr
These two emerge together but differ in name \cr
Their unity is said to be a mystery \cr
Mystery of mysteries, the door to the next revision \cr
}
\references{
Lao Tsu, Tao Te Ching.
}
\seealso{
\code{\link{do.nothing}}, \code{\link{do.nothing.else}}
}
\author{
Lao Tsu, Aaron A. King
}
\keyword{programming}
