\name{tao-package}
\docType{package}
\alias{tao-package}
\title{That which cannot be packaged.}
\description{The package that can be described is not the tao package.}
\details{
  Tao is the source of the ten thousand things.
}
\references{
  Lao Tsu, Tao Te Ching.
  Adapted from the translation by Gia-Fu Feng and Jane English, Vintage Books (a division of Random House). 
  Online at http://elyxr.com/tao.
}
\author{Aaron A. King (kingaa at umich dot edu)}
\seealso{\code{\link{do.nothing}}, \code{\link{do.nothing.else}}}
\keyword{programming}
